+++ stud_do.API/Services/PostService/PostService.cs
@@ -12,22 +12,32 @@ namespace stud_do.API.Services.PostService
             _context = context;
             _httpContextAccessor = httpContextAccessor;
         }
+
+        /// <summary>
+        /// Получение Id авторизованного пользователя
+        /// </summary>
         private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
-        public async Task<ServiceResponse<List<PostOutput>>> AddPostAsync(PostInput post)
+        public async Task<ServiceResponse<List<PostOutput>>> AddPostAsync(PostInput post, int roomId)
         {
             var result = new ServiceResponse<List<PostOutput>>();
-            if (!IsAdmin(post.RoomId) && !IsModerator(post.RoomId))
+            var dbRoom = await _context.Rooms.Where(p => p.Id == roomId).FirstOrDefaultAsync();
+            if (dbRoom == null)
             {
                 result.Success = false;
-                result.Message = "No rights";
+                result.Message = "Комната не найдена.";
+                return result;
+            }
+            if (!IsAdmin(roomId) && !IsModerator(roomId))
+            {
+                result.Success = false;
+                result.Message = "Нет прав на добавление записи.";
                 return result;
             }
             Post newPost = new()
             {
-                RoomId = post.RoomId,
-                Text = post.Text,
-                Attachments = post.Attachments
+                RoomId = roomId,
+                Text = post.Text
             };
             _context.Posts.Add(newPost);
             try
@@ -40,29 +50,70 @@ namespace stud_do.API.Services.PostService
                 result.Message = ex.Message;
                 return result;
             }
+            if (post.Attachments != null)
+            {
+                List<Attachment> atts = new();
+                foreach (var item in post.Attachments)
+                {
+                    atts.Add(new Attachment { Link = item, PostId = newPost.Id });
+                }
+                _context.Attachments.AddRange(atts);
+                try
+                {
+                    await _context.SaveChangesAsync();
+                }
+                catch (Exception ex)
+                {
+                    result.Success = false;
+                    result.Message = ex.Message;
+                    return result;
+                }
+            }
             return await GetPostsAsync(newPost.RoomId);
         }
 
+        /// <summary>
+        /// Проверка наличия прав администратора
+        /// </summary>
+        /// <param name="roomId">Id комнаты</param>
         private bool IsAdmin(int roomId)
         {
-            var room = _context.UsersRooms.Where(r => r.UserId == GetUserId() && r.RoomId == roomId).FirstOrDefaultAsync();
-            return room.Result.RoleId == 1;
+            var room = _context.UsersRooms.Where(r => r.UserId == GetUserId() && r.RoomId == roomId).FirstOrDefault();
+            if (room == null)
+            {
+                return false;
+            }
+            return room.RoleId == 1;
         }
 
+        /// <summary>
+        /// Проверка наличия прав редактора
+        /// </summary>
+        /// <param name="roomId">Id комнаты</param>
         private bool IsModerator(int roomId)
         {
-            var room = _context.UsersRooms.Where(r => r.UserId == GetUserId() && r.RoomId == roomId).FirstOrDefaultAsync();
-            return room.Result.RoleId == 2;
+            var room = _context.UsersRooms.Where(r => r.UserId == GetUserId() && r.RoomId == roomId).FirstOrDefault();
+            if (room == null)
+            {
+                return false;
+            }
+            return room.RoleId == 2;
         }
 
         public async Task<ServiceResponse<List<PostOutput>>> DeletePostAsync(int postId)
         {
             var result = new ServiceResponse<List<PostOutput>>();
             var dbPost = await _context.Posts.Where(p => p.Id == postId).FirstOrDefaultAsync();
+            if (dbPost == null)
+            {
+                result.Success = false;
+                result.Message = "Запись не найдена.";
+                return result;
+            }
             if (!IsAdmin(dbPost.RoomId) && !IsModerator(dbPost.RoomId))
             {
                 result.Success = false;
-                result.Message = "No rights";
+                result.Message = "Нет прав на удаление записи.";
                 return result;
             }
             _context.Posts.Remove(dbPost);
@@ -83,19 +134,60 @@ namespace stud_do.API.Services.PostService
         {
             var result = new ServiceResponse<PostOutput>();
             var post = await _context.Posts.Where(p => p.Id == postId).FirstOrDefaultAsync();
+            if (post == null)
+            {
+                result.Success = false;
+                result.Message = "Запись не найдена.";
+                return result;
+            }
+            var roomUsrs = await _context.UsersRooms.Where(p => p.RoomId == post.RoomId).ToListAsync();
+            if (!roomUsrs.Any(x => x.UserId == GetUserId()))
+            {
+                result.Success = false;
+                result.Message = "Нет доступа к комнате.";
+                return result;
+            }
+            var atts = await GetAttahcments(postId);
             PostOutput output = new()
             {
                 Text = post.Text,
                 RoomId = post.RoomId,
-                Id = postId
+                Id = postId,
+                Attachments = atts
             };
             result.Data = output;
             return result;
         }
 
+        /// <summary>
+        /// Получить вложения в записи
+        /// </summary>
+        /// <param name="postId">Id записи</param>
+        /// <returns></returns>
+        private async Task<List<string>> GetAttahcments(int postId)
+        {
+            var dbAtts = await _context.Attachments.Where(a => a.PostId == postId).ToListAsync();
+            var atts = new List<string>();
+            if (dbAtts != null)
+            {
+                foreach (var item in dbAtts)
+                {
+                    atts.Add(item.Link);
+                }
+            }
+            return atts;
+        }
+
         public async Task<ServiceResponse<List<PostOutput>>> GetPostsAsync(int roomId)
         {
             var result = new ServiceResponse<List<PostOutput>>();
+            var roomUsrs = await _context.UsersRooms.Where(p => p.RoomId == roomId).ToListAsync();
+            if (!roomUsrs.Any(x => x.UserId == GetUserId()))
+            {
+                result.Success = false;
+                result.Message = "Нет доступа к комнате.";
+                return result;
+            }
             var posts = await _context.Posts.Where(p => p.RoomId == roomId).ToListAsync();
             var outputs = new List<PostOutput>();
             foreach (var post in posts)
@@ -103,8 +195,9 @@ namespace stud_do.API.Services.PostService
                 PostOutput output = new()
                 {
                     Text = post.Text,
-                    RoomId = post.RoomId,
-                    Id = post.Id
+                    RoomId = roomId,
+                    Id = post.Id,
+                    Attachments = await GetAttahcments(post.Id)
                 };
                 outputs.Add(output);
             }
@@ -112,17 +205,63 @@ namespace stud_do.API.Services.PostService
             return result;
         }
 
-        public async Task<ServiceResponse<PostOutput>> UpdatePostAsync(Post post)
+        public async Task<ServiceResponse<PostOutput>> UpdatePostAsync(PostInput post, int postId)
         {
             var result = new ServiceResponse<PostOutput>();
-            if (!IsAdmin(post.RoomId) && !IsModerator(post.RoomId))
+            var dbPost = await _context.Posts.Where(p => p.Id == postId).FirstOrDefaultAsync();
+            if (dbPost == null)
             {
                 result.Success = false;
-                result.Message = "No rights";
+                result.Message = "Запись не найдена.";
                 return result;
             }
-            var dbPost = await _context.Posts.Where(p => p.Id == post.Id).FirstOrDefaultAsync();
-            dbPost = post;
+            if (!IsAdmin(dbPost.RoomId) && !IsModerator(dbPost.RoomId))
+            {
+                result.Success = false;
+                result.Message = "Нет прав на изменение записи.";
+                return result;
+            }
+            
+            dbPost.Text = post.Text;
+            var dbAtts = await _context.Attachments.Where(a => a.PostId == postId).ToListAsync();
+            if (dbAtts != null)
+            {
+                _context.Attachments.RemoveRange(dbAtts);
+                try
+                {
+                    await _context.SaveChangesAsync();
+                }
+                catch (Exception ex)
+                {
+                    result.Success = false;
+                    result.Message = ex.Message;
+                    return result;
+                }
+            }
+            if (post.Attachments != null)
+            {
+                List<Attachment> atts = new();
+                foreach (var item in post.Attachments)
+                {
+                    atts.Add(new Attachment 
+                    { 
+                        Link = item, 
+                        PostId = dbPost.Id 
+                    });
+                }
+
+                _context.Attachments.AddRange(atts);
+                try
+                {
+                    await _context.SaveChangesAsync();
+                }
+                catch (Exception ex)
+                {
+                    result.Success = false;
+                    result.Message = ex.Message;
+                    return result;
+                }
+            }
             try
             {
                 await _context.SaveChangesAsync();
@@ -136,8 +275,9 @@ namespace stud_do.API.Services.PostService
             PostOutput output = new()
             {
                 Text = post.Text,
-                RoomId = post.RoomId,
-                Id = post.Id
+                RoomId = dbPost.RoomId,
+                Id = postId,
+                Attachments = await GetAttahcments(dbPost.Id)
             };
             result.Data = output;
             return result;
